DRY ChevronLeftIcon
Extract Chevrons to their own component file
