+++ src/app/content/components/BookBanner.tsx
@@ -15,6 +15,7 @@ import * as select from '../selectors';
 import { BookWithOSWebData } from '../types';
 import { isClickWithModifierKeys } from '../utils/domUtils';
 import { bookDetailsUrl } from '../utils/urlUtils';
+import { ChevronLeftIcon } from './ChevronIcons';
 import {
   bookBannerDesktopBigHeight,
   bookBannerDesktopMiniHeight,
@@ -23,30 +24,6 @@ import {
   contentTextWidth,
 } from './constants';
 import './BookBanner.css';
-
-interface IconProps extends React.SVGAttributes<SVGSVGElement> {
-  className?: string;
-}
-
-/**
- * ChevronLeft icon for BookBanner component.
- * SVG path from Boxicons (https://boxicons.com - MIT License)
- */
-function ChevronLeftIcon({ className, ...props }: IconProps) {
-  return (
-    <svg
-      className={className}
-      viewBox="0 0 24 24"
-      aria-hidden="true"
-      {...props}
-    >
-      <path
-        fill="currentColor"
-        d="M13.293 6.293 7.586 12l5.707 5.707 1.414-1.414L10.414 12l4.293-4.293z"
-      />
-    </svg>
-  );
-}
 
 const gradients: {[key in BookWithOSWebData['theme']]: string} = {
   'blue': '#004aa2',
