+++ src/app/content/components/ChevronIcons.tsx
@@ -1,0 +1,45 @@
+import React from 'react';
+
+interface IconProps extends React.SVGAttributes<SVGSVGElement> {
+  className?: string;
+}
+
+/**
+ * ChevronLeft icon for PrevNextBar component.
+ * SVG path from Boxicons (https://boxicons.com - MIT License)
+ */
+export function ChevronLeftIcon({ className, ...props }: IconProps) {
+  return (
+    <svg
+      className={className}
+      viewBox="0 0 24 24"
+      aria-hidden="true"
+      {...props}
+    >
+      <path
+        fill="currentColor"
+        d="M13.293 6.293 7.586 12l5.707 5.707 1.414-1.414L10.414 12l4.293-4.293z"
+      />
+    </svg>
+  );
+}
+
+/**
+ * ChevronRight icon for PrevNextBar component.
+ * SVG path from Boxicons (https://boxicons.com - MIT License)
+ */
+export function ChevronRightIcon({ className, ...props }: IconProps) {
+  return (
+    <svg
+      className={className}
+      viewBox="0 0 24 24"
+      aria-hidden="true"
+      {...props}
+    >
+      <path
+        fill="currentColor"
+        d="M10.707 17.707 16.414 12l-5.707-5.707-1.414 1.414L13.586 12l-4.293 4.293z"
+      />
+    </svg>
+  );
+}
